requirements.txt
streamlit>=1.32.0
supabase>=2.3.0,<2.12.0
postgrest>=0.16.0,<1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.18.0
